Move test code to a new file
