+++ Makefile
@@ -5,7 +5,8 @@ install:
 	$(PIP) install -r requirements.txt
 
 test:
-	$(PYTHON) ssa/core.py
+	$(PIP) install .
+	$(PYTHON) test/test.py
 #	type-check
 	mypy -p ssa \
 	  --ignore-missing-imports # stubs don't exist for networkx and that's ok
